requirements.txt
djangorestframework-simplejwt==5.3.1
djangorestframework==3.15.2
psycopg2-binary==2.9.9
